Povaliaev_lesson10_3

+++ OnlineShop/OnlineShopWebApp/Services/AccountsService.cs
@@ -9,7 +9,6 @@ using OnlineShopWebApp.Interfaces;
 using OnlineShopWebApp.Models;
 using OnlineShopWebApp.Models.Abstractions;
 using System.Collections.Generic;
-using System.Data;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
@@ -20,16 +19,14 @@ namespace OnlineShopWebApp.Services
     {
         private readonly IMapper _mapper;
         private readonly RolesService _rolesService;
-        private readonly HashService _hashService;
         private readonly IExcelService _excelService;
         private readonly SignInManager<User> _signInManager;
         private readonly UserManager<User> _userManager;
 
-        public AccountsService(IMapper mapper, RolesService rolesService, HashService hashService, IExcelService excelService, SignInManager<User> signInManager, UserManager<User> userManager)
+        public AccountsService(IMapper mapper, RolesService rolesService, IExcelService excelService, SignInManager<User> signInManager, UserManager<User> userManager)
         {
             _mapper = mapper;
             _rolesService = rolesService;
-            _hashService = hashService;
             _excelService = excelService;
             _signInManager = signInManager;
             _userManager = userManager;
@@ -42,8 +39,17 @@ namespace OnlineShopWebApp.Services
         public virtual async Task<List<UserViewModel>> GetAllAsync()
         {
             var users = await _userManager.Users.ToListAsync();
-            return users.Select(_mapper.Map<UserViewModel>)
-                        .ToList();
+            var usersViewModels = new List<UserViewModel>(users.Count);
+
+            foreach (var user in users)
+            {
+                var roles = await _userManager.GetRolesAsync(user);
+                var userVM = _mapper.Map<UserViewModel>(user);
+                userVM.RoleName = roles.FirstOrDefault()!;
+                usersViewModels.Add(userVM);
+            }
+
+            return usersViewModels;
         }
 
         /// <summary>
@@ -54,7 +60,11 @@ namespace OnlineShopWebApp.Services
         public virtual async Task<UserViewModel> GetAsync(string id)
         {
             var userDb = await _userManager.FindByIdAsync(id);
-            return _mapper.Map<UserViewModel>(userDb);
+            var roles = await _userManager.GetRolesAsync(userDb!);
+            var userVM = _mapper.Map<UserViewModel>(userDb);
+            userVM.RoleName = roles.First();
+
+            return userVM;
         }
 
         /// <summary>
@@ -68,12 +78,18 @@ namespace OnlineShopWebApp.Services
                 Email = register.Email,
                 UserName = register.Email,
                 FullName = register.Name,
-                PhoneNumber = register.Phone
+                PhoneNumber = register.PhoneNumber
             };
 
             await _userManager.CreateAsync(user, register.Password);
-            await _signInManager.SignInAsync(user, false);
-            await _userManager.AddToRoleAsync(user, Constants.UserRoleName);
+
+            var roleName = await GetRegisterRoleNameAsync(register);
+            await _userManager.AddToRoleAsync(user, roleName);
+
+            if (register is not AdminRegisterViewModel)
+            {
+                await _signInManager.SignInAsync(user, false);
+            }
         }
 
         /// <summary>
@@ -90,7 +106,8 @@ namespace OnlineShopWebApp.Services
                 return;
             }
 
-            user.PasswordHash = _hashService.GenerateHash(changePassword.Password);
+            var newPasswordHash = _userManager.PasswordHasher.HashPassword(user, changePassword.Password);
+            user.PasswordHash = newPasswordHash;
 
             await _userManager.UpdateAsync(user);
         }
@@ -109,23 +126,31 @@ namespace OnlineShopWebApp.Services
                 return;
             }
 
-            var role = await _rolesService.GetAsync(editUser.RoleId);
-
             user.Email = editUser.Email;
             user.PhoneNumber = editUser.Phone;
             user.FullName = editUser.Name;
 
             await _userManager.UpdateAsync(user);
+
+            var userRoles = await _userManager.GetRolesAsync(user);
+            await _userManager.AddToRoleAsync(user, editUser.RoleName);
+            await _userManager.RemoveFromRolesAsync(user, userRoles);
         }
 
         /// <summary>
-        /// Delete user from repository by id
+        /// Delete user from repository by id. Admin can't be deleted
         /// </summary>
         /// <param name="id">Target user id (GUID)</param>
         public virtual async Task DeleteAsync(string id)
         {
             var user = await _userManager.FindByIdAsync(id);
-            await _userManager.DeleteAsync(user!);
+
+            var isAdmin = await _userManager.IsInRoleAsync(user!, Constants.AdminRoleName);
+
+            if (!isAdmin)
+            {
+                await _userManager.DeleteAsync(user!);
+            }
         }
 
         /// <summary>
@@ -164,8 +189,7 @@ namespace OnlineShopWebApp.Services
                 modelState.AddModelError(string.Empty, "Email уже зарегистрирован!");
             }
 
-
-            if (register is AdminRegisterViewModel { RoleId: var roleId } && !await IsRoleExistAsync(roleId))
+            if (register is AdminRegisterViewModel { RoleName: var roleName } && !await IsRoleExistAsync(roleName))
             {
                 modelState.AddModelError(string.Empty, "Роль не существует!");
             }
@@ -188,7 +212,7 @@ namespace OnlineShopWebApp.Services
                 modelState.AddModelError(string.Empty, "Email уже зарегистрирован!");
             }
 
-            var isRoleExist = await IsRoleExistAsync(editUser.RoleId);
+            var isRoleExist = await IsRoleExistAsync(editUser.RoleName);
 
             if (!isRoleExist)
             {
@@ -198,10 +222,10 @@ namespace OnlineShopWebApp.Services
             return modelState.IsValid;
         }
 
-        public async Task LogoutAsync()
-        {
-            await _signInManager.SignOutAsync();
-        }
+        /// <summary>
+        /// Logout user
+        /// </summary>
+        public async Task LogoutAsync() => await _signInManager.SignOutAsync();
 
         /// <summary>
         /// Change all users role related to role Id to user Role.
@@ -227,21 +251,20 @@ namespace OnlineShopWebApp.Services
         }
 
         /// <summary>
-        /// Get a role for new user based on register model
+        /// Get a role name for new user based on register model
         /// </summary>        
-        /// <returns>Associated Role Id; Return Role User Id as default</returns>
+        /// <returns>Associated Role name; Return User role name as default</returns>
         /// <param name="register">Target register model</param>
-        private async Task<string> GetRegisterRoleIdAsync(RegisterViewModel register)
+        private async Task<string> GetRegisterRoleNameAsync(RegisterViewModel register)
         {
             if (register is AdminRegisterViewModel)
             {
                 var adminRegister = register as AdminRegisterViewModel;
-                return adminRegister!.RoleId;
+                var role = await _rolesService.GetByNameAsync(adminRegister!.RoleName);
+                return role?.Name ?? Constants.UserRoleName;
             }
 
-            var roles = await _rolesService.GetAllAsync();
-            return roles.First(r => r.Name == Constants.UserRoleName)
-                        .Id;
+            return Constants.UserRoleName;
         }
 
         /// <summary>
@@ -257,15 +280,15 @@ namespace OnlineShopWebApp.Services
         }
 
         /// <summary>
-        /// Checks if a role with the given id exists.
+        /// Checks if a role with the given name exists.
         /// </summary>        
         /// <returns>true if exists; otherwise false</returns>
-        /// <param name="roleId">Target role id (GUID)</param>
-        private async Task<bool> IsRoleExistAsync(string roleId)
+        /// <param name="roleName">Target role name (GUID)</param>
+        private async Task<bool> IsRoleExistAsync(string roleName)
         {
-            var role = await _rolesService.GetAsync(roleId);
+            var role = await _rolesService.GetByNameAsync(roleName);
 
-            return role != null;
+            return role is not null;
         }
     }
 }
